TR-7 fix Blip, BlipEvent, RadarDto, RadarMapper in accordance to TR-10, TR-11 comments (equals, hashCode, small fixes)
