+++ src/main/java/ru/hh/techradar/mapper/RadarMapper.java
@@ -2,10 +2,8 @@ package ru.hh.techradar.mapper;
 
 import java.util.Collection;
 import java.util.List;
-import org.hibernate.proxy.HibernateProxy;
 import org.springframework.stereotype.Component;
 import ru.hh.techradar.dto.RadarDto;
-import ru.hh.techradar.dto.RadarShortDto;
 import ru.hh.techradar.entity.Radar;
 
 @Component
@@ -31,30 +29,20 @@ public class RadarMapper extends AbstractMapper<Radar, RadarDto> {
     RadarDto radarDto = new RadarDto();
     radarDto.setId(entity.getId());
     radarDto.setName(entity.getName());
-    if (!(entity.getQuadrants() instanceof HibernateProxy)) {
-      radarDto.setQuadrants(quadrantMapper.toDtos(entity.getQuadrants()));
-    }
-    if (!(entity.getRings() instanceof HibernateProxy)) {
-      radarDto.setRings(ringMapper.toDtos(entity.getRings()));
-    }
-    if (!(entity.getBlips() instanceof HibernateProxy)) {
-      radarDto.setBlips(blipMapper.toDtos(entity.getBlips()));
-    }
+    radarDto.setQuadrants(quadrantMapper.toDtos(entity.getQuadrants()));
+    radarDto.setRings(ringMapper.toDtos(entity.getRings()));
+    radarDto.setBlips(blipMapper.toDtos(entity.getBlips()));
     return radarDto;
   }
 
-  public RadarShortDto toShortDto(Radar entity) {
-    RadarShortDto radarShortDto = new RadarShortDto();
-    radarShortDto.setId(entity.getId());
-    radarShortDto.setName(entity.getName());
-    return radarShortDto;
+  public RadarDto toShortDto(Radar entity) {
+    RadarDto radarDto = new RadarDto();
+    radarDto.setId(entity.getId());
+    radarDto.setName(entity.getName());
+    return radarDto;
   }
 
-  public List<RadarShortDto> toShortDtos(Collection<Radar> entities) {
+  public List<RadarDto> toShortDtos(Collection<Radar> entities) {
     return entities.stream().map(this::toShortDto).toList();
-  }
-
-  public Radar toUpdate(Radar target, Radar source) {
-    return null;
   }
 }
